[level 2] Title: [PCCP 기출문제] 2번 / 퍼즐 게임 챌린지, Time: 26.04 ms, Memory: 101 MB -BaekjoonHub

+++ 프로그래머스/2/340212. ［PCCP 기출문제］ 2번 ／ 퍼즐 게임 챌린지/［PCCP 기출문제］ 2번 ／ 퍼즐 게임 챌린지.java
@@ -42,7 +42,7 @@ class Solution {
         long result = 0;
         for (int i = 0; i < diffs.length; i++) {
             if (diffs[i] <= level) result += times[i];
-            else result += ((times[i] + times[i - 1]) * (diffs[i] - level) + times[i]);
+            else result += (long) ((times[i] + times[i - 1]) * (diffs[i] - level) + times[i]);
         }
         return result;
     }
